Add CMake module for building CHIP examples with Nordic NCS
Currently Nordic platform layer of CHIP is based on nRF5
SDK. This commit is the first step needed to replace it
with nRF Connect SDK based on Zephyr RTOS. While the old
platform layer remains unchanged, add a CMake helper file
which allows to build the entire CHIP using the NCS/Zephyr
build system.

See comments in the nrf5-app.cmake for usage example.

To build the CHIP with NCS you need to:
1. install and configure NCS environment by sourcing
   zephyr-env.sh
2. export NRF5_SDK_ROOT=<path to nRF5 SDK> (still neeeded
   until the platform layer switches to NCS)
3. cd examples/lock-app/nrf5_ncs
4. west build -b nrf52840dk_nrf52840